Noam/streaming (#26)
* refactor:extract code to multiple files

* monitor stream for chat completion

* rename file

* async chat completion

* fix stream async + add spans for streams

* remove unused code

* Added stream token calculation

* fix typo + remove prints and comments

* added "stream" attribute to completion events

* handle errors before the stream starts
