+++ src/nr_openai_observability/consts.py
@@ -1,0 +1,6 @@
+EventName = "LlmCompletion"
+MessageEventName = "LlmChatCompletionMessage"
+SummaryEventName = "LlmChatCompletionSummary"
+EmbeddingEventName = "LlmEmbedding"
+VectorSearchEventName = "LlmVectorSearch"
+VectorSearchResultsEventName = "LlmVectorSearchResult"
